requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.6.3
argcomplete==2.0.0
asgiref==3.4.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.9.1
base58==2.1.0
beautifulsoup4==4.10.0
bidict==0.21.4
bitarray==1.2.2
black==21.12b0
boto3==1.17.109
botocore==1.20.109
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.2
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.4
colorthief==0.2.1
contextlib2==21.6.0
crytic-compile==0.2.4
cssselect==1.1.0
customtkinter==4.5.5
cycler==0.10.0
Cython==0.29.26
cytoolz==0.11.0
darkdetect==0.7.0
dash==1.21.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
distlib==0.3.2
Django==3.2.6
docopt==0.6.2
docutils==0.17.1
dominate==2.5.2
emoji==1.2.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fake-useragent==0.1.11
filelock==3.0.12
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Compress==1.10.1
Flask-GraphQL==2.0.1
Flask-Gravatar==0.5.0
Flask-HTTPAuth==4.7.0
Flask-Login==0.5.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==0.14.3
frozendict==2.3.4
frozenlist==1.2.0
future==0.18.2
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.5
graphql-server-core==1.2.0
greenlet==1.1.0
gunicorn==20.0.4
h11==0.14.0
hexbytes==0.2.2
httpcore==0.16.3
httpx==0.23.3
idna==2.10
importlib-metadata==4.8.1
iniconfig==1.1.1
install==1.3.4
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==3.2.0
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kiwisolver==1.3.1
lru-dict==1.1.7
lru-od==0.0.1
lxml==4.6.3
MarkupSafe==2.1.1
matplotlib==3.4.2
mediapipe==0.8.10.1
momo==0.1.0
moralis==0.1.14
MouseInfo==0.1.3
mtnmomo==3.0.1
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.25
mysqlclient==2.0.3
netaddr==0.8.0
numpy==1.21.1
opencv-contrib-python==4.6.0.66
packaging==21.3
pandas==1.3.1
parse==1.19.0
parsimonious==0.8.1
pathspec==0.9.0
patsy==0.5.1
phonenumbers==8.12.31
Pillow==8.3.1
pipenv==2021.11.15
pipreqs==0.4.10
pipx==1.0.0
platformdirs==2.4.1
plotly==5.1.0
pluggy==1.0.0
prettytable==3.4.1
promise==2.3
protobuf==3.17.3
psycopg2-binary==2.8.6
py==1.11.0
py-solc-x==1.1.1
pyaml==21.10.1
PyAutoGUI==0.9.53
pycryptodome==3.11.0
pydantic==1.8.2
pyee==8.2.2
pyfiglet==0.8.post1
pygame==2.0.2
PyGetWindow==0.0.9
Pygments==2.10.0
PyJWT==1.7.1
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
pyppeteer==0.2.6
pyquery==1.4.3
pyreadline3==3.4.1
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.28
pysha3==1.0.2
pyssml==0.1.3
pytest==6.2.5
python-dateutil==2.7.5
python-dotenv==0.19.2
python-engineio==4.3.0
python-socketio==5.5.0
python-telegram-bot==13.15
pytweening==1.0.4
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==302
PyYAML==5.4.1
requests==2.24.0
requests-html==0.10.0
rfc3986==1.5.0
rlp==2.0.1
Rx==1.6.1
s3transfer==0.4.2
schema==0.7.5
scipy==1.7.1
selenium==3.141.0
semantic-version==2.8.5
sh==1.14.2
six==1.16.0
slither-analyzer==0.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.2.1
splinter==0.18.1
SQLAlchemy==1.4.42
SQLAlchemy-serializer==1.4.1
sqlitedict==2.1.0
sqlparse==0.4.1
statsmodels==0.12.2
stripe==2.58.0
tenacity==8.0.1
tinydb==4.7.0
tkcalendar==1.6.1
toml==0.10.2
tomli==1.2.3
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
twilio==6.60.0
typing_extensions==4.3.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
userpath==1.7.0
varint==1.0.2
verlat==0.1.0.post1
virtualenv==20.4.7
virtualenv-clone==0.5.7
visitor==0.1.3
w3lib==1.22.0
wappdriver==0.6.6
wcwidth==0.2.5
web3==5.26.0
websockets==9.1
Werkzeug==2.2.2
WTForms==2.3.3
yarg==0.1.9
yarl==1.7.0
z3-solver==4.11.2.0
zipp==3.6.0
